cn0401: add application code
The application is a SPI-to-CAN translator that runs on the EVAL-ADICUP3029
with the EVAL-CN0401-ARDZ daughter board. The application sets up the
system to be a node in a CAN bus. Each node can be connected to eachother
via a CAN bus, and, in turn, to a PC via an USB cable. The connection to
the PC provides a CLI to command each node. The node can be set up to send
an ASCII message on the bus and the receiving node will wake up, receive
and display the message. Then both nodes go back to sleep.

Signed-off-by: Drimbarean <Andrei.Drimbarean@analog.com>